Implement Step 20: Full Windows Support
This commit completes the platform-specific abstractions for Windows
in `sysdeps` and refactors the ADB stack to use cross-platform `AdbFd`.

Key changes:
- Refactored `sysdeps::AdbFd` to unify Unix file descriptors and
  Windows handles/sockets.
- Implemented `Read` and `Write` for `&AdbFd` to support shared I/O
  (required for `rustls`).
- Fixed signedness mismatches in `sysdeps::errno` for Windows.
- Updated `adb-transport`, `adb-sockets`, `adb-services`, and
  `fdevent` to use `AdbFd` instead of `OwnedFd` or `TcpStream`.
- Ported test suites to be platform-agnostic using `adb_socketpair`.
- Updated GitHub Actions to run workspace tests on both Ubuntu and
  Windows.
- Updated `PORTING_PLAN.md` to reflect Step 20 completion.

+++ rust/sysdeps/src/net.rs
@@ -111,7 +111,7 @@ pub fn set_tcp_keepalive<T: AsRawSocket>(socket: &T, interval_sec: i32) -> bool 
     let enable: i32 = if interval_sec > 0 { 1 } else { 0 };
     unsafe {
         setsockopt(
-            s,
+            s as usize,
             SOL_SOCKET,
             SO_KEEPALIVE,
             &enable as *const _ as *const _,
@@ -143,7 +143,7 @@ pub fn disable_tcp_nagle<T: AsRawSocket>(socket: &T) {
     let off: i32 = 1;
     unsafe {
         setsockopt(
-            s,
+            s as usize,
             IPPROTO_TCP as _,
             TCP_NODELAY as _,
             &off as *const _ as *const _,
@@ -209,9 +209,9 @@ pub fn network_peek<T: AsRawFd>(socket: &T) -> Option<isize> {
 pub fn network_peek<T: AsRawSocket>(socket: &T) -> Option<isize> {
     ensure_wsa_startup();
     let s = socket.as_raw_socket();
-    let mut available: i32 = 0;
+    let mut available: u32 = 0;
     unsafe {
-        if ioctlsocket(s, FIONREAD, &mut available) == 0 {
+        if ioctlsocket(s as usize, FIONREAD, &mut available) == 0 {
             Some(available as isize)
         } else {
             None
